fix: Rename all the ledger folder as per Angular convention
