+++ apps/web-giddh/src/app/ledger/components/ledger-aside-pane/ledger-aside-pane.component.scss
