+++ apps/web-giddh/src/app/ledger/components/ledger-columnar-report-table/ledger-columnar-report-table.component.scss
