+++ apps/web-giddh/src/app/ledger/components/ledger-discount/ledger-discount.module.ts
@@ -4,7 +4,7 @@ import { FormsModule } from "@angular/forms";
 import { ClickOutsideModule } from "ng-click-outside";
 import { DecimalDigitsModule } from "../../../shared/helpers/directives/decimalDigits/decimalDigits.module";
 import { NgxMaskModule } from "../../../shared/helpers/directives/ngx-mask";
-import { LedgerDiscountComponent } from "./ledgerDiscount.component";
+import { LedgerDiscountComponent } from "./ledger-discount.component";
 
 @NgModule({
     declarations: [LedgerDiscountComponent],
