+++ apps/web-giddh/src/app/ledger/components/share-ledger/share-ledger.component.scss
