+++ apps/web-giddh/src/app/ledger/components/update-ledger-discount/update-ledger-discount.component.ts
@@ -1,41 +1,46 @@
 import { take, takeUntil } from 'rxjs/operators';
-import { Component, ElementRef, EventEmitter, Input, OnChanges, OnDestroy, OnInit, Output, SimpleChanges, ViewChild } from '@angular/core';
+import { Component, EventEmitter, Input, OnChanges, OnDestroy, OnInit, Output, SimpleChanges } from '@angular/core';
 import { Store, select } from '@ngrx/store';
 import { AppState } from '../../../store';
 import { Observable, ReplaySubject } from 'rxjs';
+import { INameUniqueName } from '../../../models/api-models/Inventory';
 import { IDiscountList, LedgerDiscountClass } from '../../../models/api-models/SettingsDiscount';
-import { giddhRoundOff } from '../../../shared/helpers/helperFunctions';
+
+export class UpdateLedgerDiscountData {
+    public particular: INameUniqueName = { name: '', uniqueName: '' };
+    public amount: number = 0;
+}
 
 @Component({
-    selector: 'ledger-discount',
-    templateUrl: 'ledgerDiscount.component.html',
-    styleUrls: [`./ledgerDiscount.component.scss`]
+    selector: 'update-ledger-discount',
+    templateUrl: 'update-ledger-discount.component.html',
+    styleUrls: ['./update-ledger-discount.component.scss']
 })
 
-export class LedgerDiscountComponent implements OnInit, OnDestroy, OnChanges {
-
-    public get defaultDiscount(): LedgerDiscountClass {
-        return this.discountAccountsDetails[0];
-    }
-
+export class UpdateLedgerDiscountComponent implements OnInit, OnChanges, OnDestroy {
     /* This will hold common JSON data */
     @Input() public commonLocaleData: any = {};
     @Input() public discountAccountsDetails: LedgerDiscountClass[];
     @Input() public ledgerAmount: number = 0;
-    @Output() public discountTotalUpdated: EventEmitter<{ discountTotal: number, isActive: any, discount: any }> = new EventEmitter();
+    @Output() public discountTotalUpdated: EventEmitter<number> = new EventEmitter();
     @Output() public hideOtherPopups: EventEmitter<boolean> = new EventEmitter<boolean>();
-    public discountTotal: number;
-    public discountAccountsList$: Observable<IDiscountList[]>;
-    public discountFromPer: boolean = true;
-    public discountFromVal: boolean = true;
-    public discountPercentageModal: number = 0;
-    public discountFixedValueModal: number = 0;
-    @ViewChild('disInptEle', { static: true }) public disInptEle: ElementRef;
 
     @Input() public discountMenu: boolean;
     @Input() public maskInput: string;
     @Input() public prefixInput: string;
     @Input() public suffixInput: string;
+
+    public discountTotal: number;
+    public discountAccountsList$: Observable<IDiscountList[]>;
+    public appliedDiscount: UpdateLedgerDiscountData[] = [];
+    public discountFromPer: boolean = true;
+    public discountFromVal: boolean = true;
+    public discountPercentageModal: number = 0;
+    public discountFixedValueModal: number = 0;
+
+    public get defaultDiscount(): LedgerDiscountClass {
+        return this.discountAccountsDetails[0];
+    }
 
     private destroyed$: ReplaySubject<boolean> = new ReplaySubject(1);
 
@@ -43,48 +48,32 @@ export class LedgerDiscountComponent implements OnInit, OnDestroy, OnChanges {
         this.discountAccountsList$ = this.store.pipe(select(p => p.settings.discount.discountList), takeUntil(this.destroyed$));
     }
 
-    public onFocusLastDiv(el) {
-        el.stopPropagation();
-        el.preventDefault();
-        if (!this.discountMenu) {
-            this.discountMenu = true;
-            this.hideOtherPopups.emit(true);
-            return;
-        }
-        let focussableElements = '.ledger-panel input[type=text]:not([disabled]),.ledger-panel [tabindex]:not([disabled]):not([tabindex="-1"])';
-        let focussable = Array.prototype.filter.call(document.querySelectorAll(focussableElements),
-            (element) => {
-                // check for visibility while always include the current activeElement
-                return element.offsetWidth > 0 || element.offsetHeight > 0 || element === document.activeElement;
-            });
-        let index = focussable.indexOf(document.activeElement);
-        if (index > -1) {
-            let nextElement = focussable[index + 1] || focussable[0];
-            nextElement.focus();
-        }
-        this.hideDiscountMenu();
-        return false;
-    }
-
     public ngOnInit() {
         this.prepareDiscountList();
-
-        if (this.defaultDiscount.discountType === 'FIX_AMOUNT') {
-            this.discountFixedValueModal = this.defaultDiscount.amount;
-        } else {
-            this.discountPercentageModal = this.defaultDiscount.amount;
-        }
         this.change();
     }
 
     public ngOnChanges(changes: SimpleChanges): void {
-        if ('discountAccountsDetails' in changes && changes.discountAccountsDetails.currentValue !== changes.discountAccountsDetails.previousValue || changes.ledgerAmount) {
+        if ('discountAccountsDetails' in changes && !changes.discountAccountsDetails.firstChange && changes.discountAccountsDetails.currentValue !== changes.discountAccountsDetails.previousValue) {
             this.prepareDiscountList();
 
-            if (this.defaultDiscount.discountType === 'FIX_AMOUNT') {
-                this.discountFixedValueModal = this.defaultDiscount.amount;
-            } else {
-                this.discountPercentageModal = this.defaultDiscount.amount;
+            /* check if !this.defaultDiscount.discountUniqueName so it's means
+              that this is default discount and we have added it manually not
+             from server side */
+            if (this.defaultDiscount && !this.defaultDiscount.discountUniqueName) {
+                if (this.defaultDiscount.discountType === 'FIX_AMOUNT') {
+                    this.discountFixedValueModal = this.defaultDiscount.discountValue;
+                    this.discountFromPer = false;
+                    this.discountFromVal = true;
+                } else {
+                    this.discountPercentageModal = this.defaultDiscount.discountValue;
+                    this.discountFromVal = false;
+                    this.discountFromPer = true;
+                }
+                if (!Number(this.defaultDiscount.discountValue)) {
+                    this.discountFromVal = true;
+                    this.discountFromPer = true;
+                }
             }
             this.change();
         }
@@ -99,7 +88,6 @@ export class LedgerDiscountComponent implements OnInit, OnDestroy, OnChanges {
         if (discountAccountsList && discountAccountsList.length) {
             discountAccountsList.forEach(acc => {
                 let hasItem = this.discountAccountsDetails.some(s => s.discountUniqueName === acc.uniqueName);
-
                 if (!hasItem) {
                     let obj: LedgerDiscountClass = new LedgerDiscountClass();
                     obj.amount = acc.discountValue;
@@ -116,17 +104,11 @@ export class LedgerDiscountComponent implements OnInit, OnDestroy, OnChanges {
     }
 
     public discountFromInput(type: 'FIX_AMOUNT' | 'PERCENTAGE', val: string) {
-        this.defaultDiscount.amount = parseFloat(String(val).replace(/,/g, ''));
-        this.defaultDiscount.discountValue = parseFloat(String(val).replace(/,/g, ''));
+        this.defaultDiscount.amount = parseFloat(val);
+        this.defaultDiscount.discountValue = parseFloat(val);
         this.defaultDiscount.discountType = type;
 
         this.change();
-
-        if (!val) {
-            this.discountFromVal = true;
-            this.discountFromPer = true;
-            return;
-        }
         if (type === 'PERCENTAGE') {
             this.discountFromPer = true;
             this.discountFromVal = false;
@@ -134,14 +116,19 @@ export class LedgerDiscountComponent implements OnInit, OnDestroy, OnChanges {
             this.discountFromPer = false;
             this.discountFromVal = true;
         }
+        if (!Number(val)) {
+            this.discountFromVal = true;
+            this.discountFromPer = true;
+            return;
+        }
     }
 
     /**
      * on change of discount amount
      */
-    public change(event?: any, discount?: any) {
-        this.discountTotal = giddhRoundOff(this.generateTotal(), 2);
-        this.discountTotalUpdated.emit({ discountTotal: this.discountTotal, isActive: event, discount: discount });
+    public change() {
+        this.discountTotal = Number(this.generateTotal() || 0);
+        this.discountTotalUpdated.emit(this.discountTotal);
     }
 
     /**
@@ -149,14 +136,6 @@ export class LedgerDiscountComponent implements OnInit, OnDestroy, OnChanges {
      * @returns {number}
      */
     public generateTotal(): number {
-        if (this.discountAccountsDetails && this.discountAccountsDetails[0]) {
-            if (this.discountAccountsDetails[0].amount) {
-                this.discountAccountsDetails[0].isActive = true;
-            } else {
-                this.discountAccountsDetails[0].isActive = false;
-            }
-        }
-
         let percentageListTotal = this.discountAccountsDetails.filter(f => f.isActive)
             .filter(s => s.discountType === 'PERCENTAGE')
             .reduce((pv, cv) => {
@@ -169,8 +148,8 @@ export class LedgerDiscountComponent implements OnInit, OnDestroy, OnChanges {
                 return Number(cv.discountValue) ? Number(pv) + Number(cv.discountValue) : Number(pv);
             }, 0) || 0;
 
-        let perFromAmount = ((percentageListTotal * this.ledgerAmount) / 100);
-        return perFromAmount + fixedListTotal;
+        let perFromAmount = Math.round(((percentageListTotal * (this.ledgerAmount || 0)) / 100) * 100) / 100;
+        return perFromAmount + Math.round(fixedListTotal * 100) / 100;
     }
 
     public trackByFn(index) {
@@ -181,14 +160,27 @@ export class LedgerDiscountComponent implements OnInit, OnDestroy, OnChanges {
         this.discountMenu = false;
     }
 
-    public toggleDiscountMenu() {
-        this.discountMenu = !this.discountMenu;
-    }
-
-    public discountInputBlur(event) {
-        if (event && event.relatedTarget && this.disInptEle && !this.disInptEle?.nativeElement.contains(event.relatedTarget)) {
-            this.hideDiscountMenu();
+    public onFocusLastDiv(el) {
+        el.stopPropagation();
+        el.preventDefault();
+        if (!this.discountMenu) {
+            this.discountMenu = true;
+            this.hideOtherPopups.emit(true);
+            return;
         }
+        let focussableElements = '.entrypanel input[type=text]:not([disabled]),.entrypanel [tabindex]:not([disabled]):not([tabindex="-1"])';
+        let focussable = Array.prototype.filter.call(document.querySelectorAll(focussableElements),
+            (element) => {
+                // check for visibility while always include the current activeElement
+                return element.offsetWidth > 0 || element.offsetHeight > 0 || element === document.activeElement;
+            });
+        let index = focussable.indexOf(document.activeElement);
+        if (index > -1) {
+            let nextElement = focussable[index + 1] || focussable[0];
+            nextElement.focus();
+        }
+        this.hideDiscountMenu();
+        return false;
     }
 
     public ngOnDestroy(): void {
