+++ apps/web-giddh/src/app/ledger/components/update-ledger-entry-panel/update-ledger.vm.ts
@@ -3,8 +3,8 @@ import { ILedgerTransactionItem } from '../../../models/interfaces/ledger.interf
 import { LedgerResponse } from '../../../models/api-models/Ledger';
 import { cloneDeep, filter, find, sumBy } from '../../../lodash-optimized';
 import { IFlattenAccountsResultItem } from '../../../models/interfaces/flattenAccountsResultItem.interface';
-import { UpdateLedgerTaxData } from '../updateLedger-tax-control/updateLedger-tax-control.component';
-import { UpdateLedgerDiscountComponent } from '../updateLedgerDiscount/updateLedgerDiscount.component';
+import { UpdateLedgerTaxData } from '../update-ledger-tax-control/update-ledger-tax-control.component';
+import { UpdateLedgerDiscountComponent } from '../update-ledger-discount/update-ledger-discount.component';
 import { TaxControlData } from '../../../theme/tax-control/tax-control.component';
 import { IOption } from '../../../theme/ng-virtual-select/sh-options.interface';
 import { LedgerDiscountClass } from '../../../models/api-models/SettingsDiscount';
