+++ apps/web-giddh/src/app/ledger/components/update-ledger-tax-control/update-ledger-tax-control.component.ts
@@ -34,8 +34,8 @@ export class UpdateLedgerTaxData {
 
 @Component({
     selector: 'update-ledger-tax-control',
-    templateUrl: 'updateLedger-tax-control.component.html',
-    styleUrls: [`./updateLedger-tax-control.component.scss`],
+    templateUrl: 'update-ledger-tax-control.component.html',
+    styleUrls: [`./update-ledger-tax-control.component.scss`],
     providers: [TAX_CONTROL_VALUE_ACCESSOR]
 })
 export class UpdateLedgerTaxControlComponent implements OnDestroy, OnChanges {
@@ -81,7 +81,7 @@ export class UpdateLedgerTaxControlComponent implements OnDestroy, OnChanges {
     private selectedTaxes: UpdateLedgerTaxData[] = [];
 
     constructor() {
-        
+
     }
 
     public ngOnChanges(changes: SimpleChanges) {
