+++ apps/web-giddh/src/app/shared/shared.module.ts
@@ -69,7 +69,7 @@ import localeMr from '@angular/common/locales/mr';
 import { ValidateSectionPermissionDirectiveModule } from './validate-section-permission/validate-section-permission.module';
 import { AmountFieldComponentModule } from './amount-field/amount-field.module';
 import { AccountAddNewDetailsModule } from './header/components/account-add-new-details/account-add-new-details.module';
-import { LedgerDiscountModule } from '../ledger/components/ledgerDiscount/ledgerDiscount.module';
+import { LedgerDiscountModule } from '../ledger/components/ledger-discount/ledger-discount.module';
 import { ConfirmationModalModule } from '../common/confirmation-modal/confirmation-modal.module';
 import { DatepickerWrapperModule } from './datepicker-wrapper/datepicker.wrapper.module';
 import { HamburgerMenuModule } from './header/components/hamburger-menu/hamburger-menu.module';
